Complete Parser Overhaul
Complete overhaul of the parser. It is far less convoluted and works far better. It is still incomplete, though it is well on its way.
The major change was in the parsing of binary expressions; there is a far more efficient algorithm in place now to do it. The rest were trimmed down as well due to how binary expressions (and expressions in general) are now parsed.