Added Pointer Support
The SIN language now supports pointers! In order to do this, however, the structure of the interpreter needed to be changed -- the symbol table type was switched from std::vector to std::list; this is because pointers to vector items are invalidated if items are added to or removed from the vector, while this is not the case for a list.
A test file was added to demonstrate pointer syntax and capability. Double and triple ref pointers are possible in SIN.
The Lexer and Parser were also updated to properly support pointer syntax.
Another small edit that was made was in the way the parser and lexer handle comment lines -- there were bugs before that are now fixed, and all comment lines are ignored.