Began work on VM and SINASM
Began work on the VM for which SIN code will be compiled. The pure interpreter still works, but the focus now will be on getting the VM to a working state. Currently, not all instructions are supported but the framework exists. the .sina filetype is now supported, and the program will assemble and execute them. Support for .sinc files is coming shortly.