Added syscall instruction and bitshifting
The program now supports, to a limited extent, bitshifting and the 'syscall' instruction, which serves as the bridge between the VM and the host. There were also more edits made including moving routines into functions; eliminating the 'inputb' and 'outputb' instructions; using regex for instruction matching so that lowercase mnemonics may be used;  more comments added.