Use BinaryIO library
Minor fix - instead of "FundamentalDataTypes", which was originall written for use in the 3.5 VCS program, this program will use BinaryIO for use in the .sinc format, which is essentially the original library under a different name. However, if I ever update BinaryIO, I will more easily be able to update the files in this project.