Type System Overhaul
The Type system was overhauled to allow for subtypes, the Interpreter symbol table was modified to use typedef struct instead of tuples, and support for indexed indirect addressing was added to the assembler. Some unnecessary/unused/old code was removed, and some comments and messages were modified to more accurately reflect the code.
See "commit_history.txt" for further details