Parser, Assembler, and Binary Tree Fixes
Fixed many bugs in the parser in addition to major code refactoring (splitting the routine up into many functions). Indirect addressing modes added to assembler. Binary trees in compiler now support signed multiplication. A lot was implemented and fixed in regards to pointers, but there is still a long way to go.
See the commit history document for more information.