Implemented rudimentary string concatenation support
There are still many bugs in the string concatenation method. I intend on changing the algorithm for binary tree evaluation and seeing if that solves the issue.
Heap allocation now uses the list::insert method instead of the list::push_back and list::sort methods -- this reduces the complexity of allocating memory in the VM.
More bug fixes were made, specifically in heap allocation.