Modified binary tree evaluation, fixed string concatenation
Binary trees are now evaluated by looking at the left-most operand in the tree rather than the right; this allows the expression to be done left-to-right. It didn't much matter before (due to the way data could be moved around), but once string concatenation was implemented, doing it right-to-left made it impossible to concatenate more than three strings at a time due to the way the strings are concatenated together (by writing to a buffer).  Now that the algorithm has been modified, string concatenation issues have been fixed.