Added function return statements
You can now use function calls as expressions. The return statement is _required_ for all functions, so if the type is void, you may say `return void` or `return`. The return statement is responsible for unwinding the stack, so it is necessary to avoid data corruption.
Fixed logical bugs in the SINVM relating to the mult and div instructions.