Code reorganization and code refactoring
Thie biggest change in this commit is that all of the code was reorganized into separate folders, and some code implementation (namely in the Compiler class) was split across multiple files. This was done for readability purposes.
Further, bugs were fixed in binary tree evaluation and in function definitions and calls; previously, it was not properly handling and tracking stack position, so any code that used the stack failed horribly upon being compiled and executed in the VM. The compiler now properly tracks and modifies stack offsets (including max offsets and where stack frames begin/end) so that this issue can be avoided.
See the commit history document for a mode detailed commit message.