Added ALU class
The VM's ADDCA, SUBCA, MULTA/MULTUA, DIVA/DIVUA instructions now use an "ALU" object. This will be the class that handles all ALU-related operations.
Further, these instructions have all been modified in how they are executed to use the STATUS register appropriately, or are in that process. Because of this, code generation in the compiler had to be modified.
Additionally, the implementation of the SINVM is in the process of being split across multiple files for organization. This process is far from complete, but the SYSCALL instruction is now located within its own .cpp file.